requirements.txt
requests>2.7
geojson_rewind
geomet
appdirs>=1.4.3
matplotlib
sphinx_issues
requests-cache
shapely>=1.5.13
vcrpy
pytest
pytest-cov
codecov
